requirements.txt
tldextract
beautifulsoup4
python-dotenv
requests